+++ lib/appium_lib/android/espresso/element/button.rb
@@ -63,9 +63,9 @@ module Appium
           # uiautomator index doesn't support last
           # and it's 0 indexed
           button_index = tags(::Appium::Android::BUTTON).length
-          button_index -= 1 if button_index > 0
+          button_index -= 1 if button_index.positive?
           image_button_index = tags(::Appium::Android::IMAGE_BUTTON).length
-          image_button_index -= 1 if image_button_index > 0
+          image_button_index -= 1 if image_button_index.positive?
 
           _button_visible_selectors_xpath(button_index: button_index,
                                           image_button_index: image_button_index)