+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Commit based release not is [release_notes.md](./release_notes.md)
 Release tags are https://github.com/appium/ruby_lib/releases .
 
 ## Unreleased
+
+Bump supported Ruby version to 2.4+
+
+[related changes](https://github.com/appium/ruby_lib_core/blob/master/CHANGELOG.md#400---2020-12-19) in ruby_lib_core.
+
 ### 1. Enhancements
 
 ### 2. Bug fixes